Refactor integration tests and update service registry configuration
- Removed CorrelationIdFilterIntegrationTest and replaced it with ObservabilityWebFilterIntegrationTest to enhance observability features.
- Updated AbstractIntegrationTest to use the new ObservabilityWebFilterIntegrationTest for logging and correlation tests.
- Added new JSON mapping files for observability tests.
- Deferred native support for the service-registry module due to AOT code generation issues with Eureka's RefreshScope.
- Updated docker-compose.yml for config-server with improved memory limits and added security environment variables.
