--- api-gateway/src/main/java/com/example/api/gateway/filter/CorrelationIdGatewayFilterFactory.java
@@ -1,50 +1,0 @@
-/***
-<p>
-    Licensed under MIT License Copyright (c) 2025-2026 Raja Kolli.
-</p>
-***/
-
-package com.example.api.gateway.filter;
-
-import java.util.UUID;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.cloud.gateway.filter.GatewayFilter;
-import org.springframework.cloud.gateway.filter.factory.AbstractGatewayFilterFactory;
-import org.springframework.core.annotation.Order;
-import org.springframework.stereotype.Component;
-
-@Component
-@Order(1)
-public class CorrelationIdGatewayFilterFactory extends AbstractGatewayFilterFactory<Object> {
-
-    private static final Logger log =
-            LoggerFactory.getLogger(CorrelationIdGatewayFilterFactory.class);
-    private static final String CORRELATION_ID_HEADER = "X-Correlation-ID";
-
-    @Override
-    public GatewayFilter apply(Object config) {
-        return (exchange, chain) -> {
-            String correlationId =
-                    exchange.getRequest().getHeaders().getFirst(CORRELATION_ID_HEADER);
-
-            if (correlationId == null || correlationId.isEmpty()) {
-                correlationId = UUID.randomUUID().toString();
-                log.debug("Generated new correlation ID: {}", correlationId);
-            } else {
-                log.debug("Using existing correlation ID: {}", correlationId);
-            }
-
-            // Add correlation ID to request
-            var request =
-                    exchange.getRequest()
-                            .mutate()
-                            .header(CORRELATION_ID_HEADER, correlationId)
-                            .build();
-
-            // Add correlation ID to response
-            exchange.getResponse().getHeaders().add(CORRELATION_ID_HEADER, correlationId);
-            return chain.filter(exchange.mutate().request(request).build());
-        };
-    }
-}
