--- api-gateway/src/test/java/com/example/api/gateway/filter/LoggingFilterIntegrationTest.java
@@ -1,64 +1,0 @@
-/***
-<p>
-    Licensed under MIT License Copyright (c) 2025 Raja Kolli.
-</p>
-***/
-
-package com.example.api.gateway.filter;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import com.example.api.gateway.config.AbstractIntegrationTest;
-import org.junit.jupiter.api.Test;
-import org.springframework.http.MediaType;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-
-public class LoggingFilterIntegrationTest extends AbstractIntegrationTest {
-
-    @DynamicPropertySource
-    static void configureProperties(DynamicPropertyRegistry registry) {
-        registry.add(
-                "spring.cloud.gateway.server.webflux.routes[0].uri", wireMockServer::getBaseUrl);
-        registry.add("spring.cloud.gateway.server.webflux.routes[0].id", () -> "logging-test");
-        registry.add(
-                "spring.cloud.gateway.server.webflux.routes[0].predicates[0]",
-                () -> "Path=/test/**");
-    }
-
-    @Test
-    void shouldLogRequestPath() {
-        webTestClient
-                .get()
-                .uri("/test/endpoint")
-                .accept(MediaType.APPLICATION_JSON)
-                .exchange()
-                .expectStatus()
-                .isOk()
-                .expectHeader()
-                .exists("X-Trace-Id")
-                .expectBody(String.class)
-                .consumeWith(
-                        result -> {
-                            assertThat(result.getResponseBody()).contains("test response");
-                        });
-    }
-
-    @Test
-    void shouldLogActuatorRequestPathAsTrace() {
-        webTestClient
-                .get()
-                .uri("/actuator/health")
-                .accept(MediaType.APPLICATION_JSON)
-                .exchange()
-                .expectStatus()
-                .isOk()
-                .expectHeader()
-                .exists("X-Trace-Id")
-                .expectBody(String.class)
-                .consumeWith(
-                        result -> {
-                            assertThat(result.getResponseBody()).contains("UP");
-                        });
-    }
-}
