--- api-gateway/src/test/resources/com/example/api/gateway/filter/LoggingFilterIntegrationTest/logging-test.json
@@ -1,19 +1,0 @@
-{
-    "mappings": [
-        {
-            "request": {
-                "method": "GET",
-                "urlPattern": "/test/endpoint"
-            },
-            "response": {
-                "status": 200,
-                "headers": {
-                    "Content-Type": "application/json"
-                },
-                "jsonBody": {
-                    "message": "test response"
-                }
-            }
-        }
-    ]
-}
