+++ api-gateway/src/test/resources/com/example/api/gateway/filter/ObservabilityWebFilterIntegrationTest/logging-test.json
@@ -1,0 +1,19 @@
+{
+    "mappings": [
+        {
+            "request": {
+                "method": "GET",
+                "urlPattern": "/test/endpoint"
+            },
+            "response": {
+                "status": 200,
+                "headers": {
+                    "Content-Type": "application/json"
+                },
+                "jsonBody": {
+                    "message": "test response"
+                }
+            }
+        }
+    ]
+}
