+++ service-registry/docker-compose.yml
@@ -2,17 +2,21 @@ version: "3.9"
 services:
   config-server:
     container_name: config-server
-    image: dockertmt/mmv2-config-server-25:0.0.1-SNAPSHOT
+    image: dockertmt/mmv2-config-server:0.0.1-SNAPSHOT
+    extra_hosts: [ 'host.docker.internal:host-gateway' ]
     deploy:
       resources:
         limits:
-          memory: 700m
+          memory: 1400m
     ports:
       - "8888:8888"
     networks:
       - microservice-network
     environment:
-      SPRING_PROFILES_ACTIVE: native
+      - SPRING_PROFILES_ACTIVE=native
+      - SPRING_SECURITY_USER_NAME=docker-dev
+      - SPRING_SECURITY_USER_PASSWORD=docker-pass
+      - JAVA_TOOL_OPTIONS=-XX:MaxDirectMemorySize=64M
 
 networks:
   microservice-network:
